trimmed down histogram
