+++ src/ndv/views/__init__.py
@@ -9,7 +9,7 @@ from typing import TYPE_CHECKING, cast
 if TYPE_CHECKING:
     from typing import Literal, TypeAlias
 
-    from ndv.views.protocols import PCanvas, PHistogramCanvas, PHistogramView, PView
+    from ndv.views.protocols import PCanvas, PHistogramCanvas, PView
 
     GuiFrontend: TypeAlias = Literal["qt", "jupyter"]
     CanvasBackend: TypeAlias = Literal["vispy", "pygfx"]
@@ -30,22 +30,6 @@ def get_view_frontend_class() -> type[PView]:
         from ._qt.qt_view import QtViewerView
 
         return QtViewerView
-
-    raise RuntimeError("No GUI frontend found")
-
-
-# TODO: add a way to set the frontend via an environment variable
-# (for example, it should be possible to use qt frontend in a jupyter notebook)
-def get_histogram_frontend_class() -> type[PHistogramView]:
-    frontend = _determine_gui_frontend()
-    if frontend == "jupyter":
-        from ._jupyter.jupyter_view import JupyterHistogramView
-
-        return JupyterHistogramView
-    if frontend == "qt":
-        from ._qt.qt_view import QtHistogramView
-
-        return QtHistogramView
 
     raise RuntimeError("No GUI frontend found")
 
@@ -74,13 +58,13 @@ def get_canvas_class(backend: str | None = None) -> type[PCanvas]:
 def get_histogram_canvas_class(backend: str | None = None) -> type[PHistogramCanvas]:
     backend = backend or os.getenv("NDV_CANVAS_BACKEND", None)
     if backend == "vispy" or (backend is None and "vispy" in sys.modules):
-        from ndv.views._vispy._vispy import VispyHistogramCanvas
+        from ndv.views._vispy._histogram import VispyHistogramCanvas
 
         return VispyHistogramCanvas
 
     if backend is None:
         if importlib.util.find_spec("vispy") is not None:
-            from ndv.views._vispy._vispy import VispyHistogramCanvas
+            from ndv.views._vispy._histogram import VispyHistogramCanvas
 
             return VispyHistogramCanvas
