+++ src/ndv/views/_jupyter/jupyter_view.py
@@ -9,7 +9,6 @@ from psygnal import Signal
 
 from ndv._types import MouseMoveEvent
 from ndv.models._array_display_model import ChannelMode
-from ndv.views.protocols import CursorType, PHistogramCanvas
 
 if TYPE_CHECKING:
     from collections.abc import Container, Hashable, Mapping, Sequence
@@ -274,85 +273,3 @@ class JupyterViewerView:
     def _on_channel_mode_changed(self, change: dict[str, Any]) -> None:
         """Emit signal when the channel mode changes."""
         self.channelModeChanged.emit(ChannelMode(change["new"]))
-
-
-class JupyterHistogramView:
-    """A Jupyter wrapper around a 'backend' Histogram View."""
-
-    def __init__(self, backend_widget: PHistogramCanvas, **kwargs: Any) -> None:
-        self._backend = backend_widget
-        # TODO: Rename
-        self._canvas = backend_widget.widget()
-        self._mouse_down: bool = False
-
-        # patch the handle_event from _jupyter_rfb.CanvasBackend
-        # to intercept various mouse events.
-        if hasattr(self._canvas, "handle_event"):
-            self._original_handle_event = self._canvas.handle_event
-            self._canvas.handle_event = self.handle_event
-
-        self._log = widgets.ToggleButton(
-            value=False,
-            description="Logarithmic Range",
-            button_style="",  # 'success', 'info', 'warning', 'danger' or ''
-            tooltip="Toggle logarithmic (base-10) range scaling",
-        )
-        self._log.observe(self._on_log_changed, names="value")
-
-        # TODO: An icon like the Qt version would be nice
-        self._reset = widgets.Button(
-            description="Reset Pan/Zoom",
-            button_style="",  # 'success', 'info', 'warning', 'danger' or ''
-            # TODO: Workshop tooltip
-            tooltip="Resets Pan and Zoom to the extent of canvas components",
-        )
-        self._reset.on_click(self._resetZoom)
-        self._btns = widgets.HBox([self._log, self._reset])
-        self.layout = widgets.VBox([self._canvas, self._btns])
-
-    def show(self) -> None:
-        """Show the viewer."""
-        from IPython.display import display
-
-        display(self.layout)  # type: ignore [no-untyped-call]
-
-    def handle_event(self, ev: dict) -> None:
-        etype = ev["event_type"]
-        intercepted = False
-        if etype == "pointer_down":
-            pos = (ev["x"], ev["y"])
-            intercepted |= self._backend.on_mouse_press(pos)
-            self._mouse_down = True
-        if etype == "pointer_move":
-            pos = (ev["x"], ev["y"])
-            intercepted |= self._backend.on_mouse_move(pos)
-            if not self._mouse_down:
-                self._set_cursor(self._backend.get_cursor(pos))
-        if etype == "pointer_up":
-            pos = (ev["x"], ev["y"])
-            intercepted |= self._backend.on_mouse_release(pos)
-            self._mouse_down = False
-
-        if not intercepted:
-            self._original_handle_event(ev)
-
-    def refresh(self) -> None:
-        self._backend.refresh()
-
-    def _resetZoom(self, wdg: widgets.Button) -> None:
-        self._backend.set_domain(None)
-        self._backend.set_range(None)
-
-    def _on_log_changed(self, change: dict[str, Any]) -> None:
-        self._backend.set_range_log(self._log.value)
-
-    def _set_cursor(self, type: CursorType) -> None:
-        # FIXME: mypy errors
-        if type is CursorType.DEFAULT:
-            self._canvas.cursor = "default"
-        elif type is CursorType.V_ARROW:
-            self._canvas.cursor = "ns-resize"
-        elif type is CursorType.H_ARROW:
-            self._canvas.cursor = "ew-resize"
-        elif type is CursorType.ALL_ARROW:
-            self._canvas.cursor = "move"
