+++ src/ndv/views/protocols.py
@@ -189,19 +189,6 @@ class PHistogramCanvas(PLutView, Protocol):
         """
 
 
-class PHistogramView(Protocol):
-    """Frontend object viewing a PHistogramCanvas."""
-
-    def __init__(self, histogram_widget: PHistogramCanvas, **kwargs: Any) -> None: ...
-    def show(self) -> None: ...
-    def set_name(self, name: str) -> None: ...
-    def set_auto_scale(self, auto: bool) -> None: ...
-    def set_colormap(self, cmap: cmap.Colormap) -> None: ...
-    def set_clims(self, clims: tuple[float, float]) -> None: ...
-    def set_lut_visible(self, visible: bool) -> None: ...
-    def setVisible(self, visible: bool) -> None: ...
-
-
 class CanvasElement(Protocol):
     """Protocol defining an interactive element on the Canvas."""
